Invert connector core provider primitive ownership
